requirements.txt
aiohttp==3.7.4.post0
asgiref==3.3.4
async-timeout==3.0.1
attrs==21.2.0
Authlib==0.15.3
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cryptography==3.4.7
Django==3.2
django-admin-rangefilter==0.8.1
django-environ==0.4.5
django-js-asset==1.2.2
django-modeltranslation==0.17.1
django-user-agents==0.4.0
djangorestframework==3.12.4
geoip2==4.1.0
gunicorn==20.1.0
h11==0.12.0
idna==2.10
ipaddress==1.0.23
ipdata==3.4.3
maxminddb==2.0.3
multidict==5.1.0
Pillow==8.2.0
psycopg2==2.8.6
pycparser==2.20
pytz==2021.1
requests==2.25.1
simple-salesforce==1.11.1
six==1.15.0
sqlparse==0.4.1
tqdm==4.60.0
typing-extensions==3.10.0.0
ua-parser==0.10.0
urllib3==1.26.4
user-agents==2.2.0
uvicorn==0.13.4
whitenoise==5.2.0
yarl==1.6.3
